+++ src/routes/router.jsx
@@ -1,9 +1,8 @@
 import App from '@/App.jsx';
-import ErrorBoundary from '@/components/feedback/ErrorBoundary.jsx';
-import PendingUI from '@/components/feedback/PendingUI.jsx';
+import { ErrorBoundary, PendingUI } from '@/components/feedback';
 import { ENDPOINTS } from '@/constants/api';
-import NotFoundPage from '@/pages/error/NotFoundPage';
-import fetchData from '@/utils/fetchData';
+import { NotFoundPage } from '@/pages/error';
+import { fetchData } from '@/utils';
 import { lazy } from 'react';
 import { createBrowserRouter } from 'react-router-dom';
